pyproject.toml
[tool.poetry]
keywords = [
  "tides",
  "marsh",
  "tidal flat",
  "mud flat",
  "marsh elevation",
  "marsh evolution",
  "marsh dynamics",
  "marsh modeling",
]
name = "tidal_marsh"
version = "0.1.0"
description = "A zero-dimensional morphodynamic model of tidal marsh elevation."
authors = ["Chris Tasich <chris.tasich@vanderbilt.edu>"]
license = "MIT"
repository = "https://gitlab.jgilligan.org/chris/tidal_marsh"
classifiers = [
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: GIS",
  "Topic :: Scientific/Engineering :: Hydrology",
  "Topic :: Scientific/Engineering :: Mathematics",
  "Topic :: Scientific/Engineering :: Physics",
  "Topic :: Scientific/Engineering :: Visualization",
  "Topic :: Scientific/Engineering :: Information Analysis",
]

[tool.poetry.group.dev.dependencies]
loguru = "^0.6.0"
ipywidgets = "^7"
flake8 = "^6.0.0"
pre-commit = "^2.20.0"
black = "^22.12.0"
mypy = "^0.991"
ipympl = "^0.9.2"
memory-profiler = "^0.61.0"
autopep8 = "^2.0.0"
line-profiler = "^4.0.1"
skyfield = "^1.45"
tqdm-batch = "^0.1.0"
psutil = "^5.9.0"
joblib = "^1.1.0"
scikit-image = "^0.19.3"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.11"
numpy = "^1.23.3"
pandas = "^1.5.2"
seaborn = "^0.12.1"
tqdm = "^4.64.1"
scipy = "^1.9.1"
matplotlib = "^3.5.3"
statsmodels = "^0.13.2"
UTide = "^0.3.0"
scikit-learn = "^1.1.3"
hydra-core = "^1.3.1"
cmocean = "^2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = false
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

#[tool.setuptools]
#package-dir = { "" = "src" }

#[tool.setuptools.packages.find]
#where = ["src"]
